+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
  - Removed the deprecated `index_buffer` module. Use the `index` module instead.
  - Fixed viewport dimensions on retina screens.
+ - The `Backend` trait is now marked unsafe.
 
 ## Version 0.1.3 (2015-03-24) & 0.1.4 (2015-03-26)